feature initdelete
