+++ src/components/index.js
@@ -3,10 +3,10 @@ export * from './inits/initDeleteCompleted.js'
 export * from './inits/initAddTodo.js'
 export * from './inits/initDownload.js'
 export * from './inits/initChangeStatus.js'
-// export * from './inits/initDelete.js'
+export * from './inits/initDelete.js'
 
 export * from './ui/updateTasksOrder.js'
 export * from './ui/addNewTodo.js'
 export * from './ui/updateTask.js'
-// export * from './ui/renderData.js'
-export * from './ui/loadData.js'
+export * from './ui/renderData.js'
+// export * from './ui/loadData.js'
