+++ src/components/inits/initChangeStatus.js
@@ -1,6 +1,5 @@
 import { toggleTodoStatus } from '../../API/index.js'
-import { loadData } from '../index.js'
-// import { loadData } from '../index.js'
+import { loadData } from '../../app.js'
 import { showError } from '../../utils/helpers.js'
 
 export function initChangeStatus(todo, checkbox) {
