+++ src/components/inits/initDeleteCompleted.js
@@ -1,7 +1,7 @@
 import { showError } from '../../utils/helpers.js'
 import { deleteCompletedTodos } from '../../API/index.js'
-// import { deleteCompletedButton, loadData, container } from '../index.js'
-import { deleteCompletedButton, loadData, container } from '../../app.js'
+import { deleteCompletedButton } from '../index.js'
+import { loadData, container } from '../../app.js'
 
 export function initDeleteCompleted() {
   deleteCompletedButton.addEventListener('click', async () => {
