+++ src/components/inits/initDownload.js
@@ -1,6 +1,7 @@
-import { loadData } from '../index.js'
+import { loadData } from '../../app.js'
 
 export const downloadButton = document.querySelector('.button-download')
+
 export function initDownload() {
   downloadButton.addEventListener('click', loadData)
 }
